+++ app/lib/shared/versions.dart
@@ -54,9 +54,9 @@ bool shouldGCVersion(String version) =>
 
 // keep in-sync with SDK version in .mono_repo.yml and Dockerfile
 final String runtimeSdkVersion = '2.19.6';
-final String toolStableDartSdkVersion = '2.19.6';
-final String toolStableFlutterSdkVersion = '3.7.12';
-final String toolPreviewDartSdkVersion = '3.0.0-417.3.beta';
+final String toolStableDartSdkVersion = '3.0.0-417.4.beta';
+final String toolStableFlutterSdkVersion = '3.10.0-1.4.pre';
+final String toolPreviewDartSdkVersion = '3.0.0-417.4.beta';
 final String toolPreviewFlutterSdkVersion = '3.10.0-1.4.pre';
 
 final semanticToolStableDartSdkVersion =