+++ CHANGELOG.md
@@ -3,6 +3,9 @@ AppEngine version, listed here to ease deployment and troubleshooting.
 
 ## Next Release (replace with git tag when deployed)
  * Bumped runtimeVersion to `2023.05.02`.
+ * Upgraded stable Dart analysis SDK to `3.0.0-417.4.beta`.
+ * Upgraded preview Dart analysis SDK to `3.0.0-417.4.beta`.
+ * Upgraded stable Flutter analysis SDK to `3.10.0-1.4.pre`.
  * Upgraded dartdoc to `6.2.2`.
 
 ## `20230427t134200-all`